gsima-agent/requirements.txt
gymnasium[other]
minigrid
mlx
mlx-lm
mlx-vlm
python-dotenv
requests
pytest
numpy
